// frontend/client/lib/types.ts
export type FoodCategory = {
    _id: string;
    categoryName: string;
  };
  
  export type Food = {
    _id: string;
    name: string;
    description: string;
    price: number;
    image?: string;
    category: string;
    quantity?: number;
  };
  
  export type OrderType = {
    _id: string;
    totalPrice: number;
    status: string;
    foodOrderItems: {
      food: {
        name: string;
        price: number;
        image: string;
      };
      quantity: number;
    }[];
  };
  
  export type CartType = {
    [_key: string]: {
      quantity: number;
    } & Food;
  };